Adjust overlay Popup and PopupFeature
